Move payment strategies into separate package paymentstrategy
