+++ example/android/SNetDemo/app/src/main/java/com/example/snetdemo/StyleTransferActivity.java
@@ -30,9 +30,9 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 
-import io.singularitynet.sdk.client.OnDemandPaymentChannelPaymentStrategy;
 import io.singularitynet.sdk.client.PaymentStrategy;
 import io.singularitynet.sdk.client.ServiceClient;
+import io.singularitynet.sdk.paymentstrategy.OnDemandPaymentChannelPaymentStrategy;
 import io.singularitynet.service.styletransfer.StyleTransferGrpc;
 import io.singularitynet.service.styletransfer.StyleTransferOuterClass;
