+++ sdk/src/main/java/io/singularitynet/sdk/paymentstrategy/CombinedPaymentStrategy.java
@@ -1,10 +1,13 @@
-package io.singularitynet.sdk.client;
+package io.singularitynet.sdk.paymentstrategy;
 
 import lombok.ToString;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import io.singularitynet.sdk.payment.Payment;
+import io.singularitynet.sdk.client.PaymentStrategy;
+import io.singularitynet.sdk.client.ServiceClient;
+import io.singularitynet.sdk.client.GrpcCallParameters;
 
 /**
  * Payment strategy which combines a list of the underlying strategies. It gets
